Refine VFX library
1.  Move all docuement specific sections to separate files
 - Move VKGC pipeline specific sections to  vfxVkgcSection.cpp/.h
 - Move render document specific section to vkRenderSection.cpp/.h
 - Add macro VFX_SUPPORT_VKGC_PIPELINE and VFX_SUPPORT_RENDER_DOCOUMENT to isolate related code.

2. Simplify section type, store shader stage in section property instead of section type

3. merge class VfxParser and  Document, it is to allow derived document object to create document specific sections.
